refactor: quality gate — split sync_tests past the 1000-line cap, shorten apply_sync_presentation
Two mechanical fixes for the Quality Checks failures on the round-3
delta, no behavior change (same 94 persistence tests, all green):

- sync_tests.rs (1105 lines, hard cap 1000) is split: the
  trash/tombstone/carryover subset (Decision A/B, S2/S4/S7/R3 trash
  semantics — 9 tests) moves verbatim to sync_trash_tests.rs; the
  shared helpers (repo/slide/row/updated_at_of/peer_song) move to
  sync_test_support.rs, used by both. Now 622 + 452 + 57 lines.
- apply_sync_presentation (123 lines, hard cap 120) extracts step 3
  into apply_unknown_sync_id() — the tombstone-aware create-or-skip
  branch (unknown + within PRUNE_HORIZON creates, older skips). Down
  to 100 lines; commit/transaction structure unchanged.

quality-check.sh --strict --against origin/main now exits 0 locally.
